Read This.
When trying to merge branches things got really messy and nothing was
working right.

If you are working on mythic make sure you ALWAYS have the latest from
the Dev-Branch and you communicate so 2 ppl don't mess with the same
things.

If you want to use this for an auto-updating server use the Jam-Craft
branch. I do my best to not push anything there that could break such.